Move MVVM Toolkit internals tests to separate project
